backend/requirements.txt

flask==2.3.3
flask-cors==4.0.0
mysql-connector-python==8.2.0
pandas==2.1.1
numpy==1.26.0
scikit-learn==1.3.1
python-dotenv==1.0.0
